refactored tests, improved typing names
