+++ packages/material-ui/src/Button/Button.test.js
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import PropTypes from 'prop-types';
 import { expect } from 'chai';
 import {
   getClasses,
@@ -7,11 +6,9 @@ import {
   describeConformance,
   act,
   createClientRender,
-  screen,
   fireEvent,
   createServerRender,
 } from 'test/utils';
-import { ThemeProvider, createMuiTheme } from '../styles';
 import Button from './Button';
 import ButtonBase from '../ButtonBase';
 
@@ -368,152 +365,6 @@ describe('<Button />', () => {
     it('should server-side render', () => {
       const markup = serverRender(<Button>Hello World</Button>);
       expect(markup.text()).to.equal('Hello World');
-    });
-  });
-
-  describe('custom theme variants', () => {
-    const WrappedComponent = (props) => {
-      const { theme, ...other } = props;
-      return (
-        <ThemeProvider theme={theme}>
-          <Button data-testid="component" {...other}>
-            Content
-          </Button>
-        </ThemeProvider>
-      );
-    };
-
-    WrappedComponent.propTypes = {
-      theme: PropTypes.object,
-    };
-
-    it('should map the variant classkey to the component', function test() {
-      if (/jsdom/.test(window.navigator.userAgent)) {
-        // see https://github.com/jsdom/jsdom/issues/2953
-        this.skip();
-      }
-
-      const theme = createMuiTheme({
-        variants: {
-          MuiButton: [
-            {
-              props: { variant: 'test' },
-              styles: { backgroundColor: 'rgb(255, 0, 0)' },
-            },
-          ],
-        },
-      });
-
-      render(<WrappedComponent theme={theme} variant="test" />);
-
-      const style = window.getComputedStyle(screen.getByTestId('component'));
-      expect(style.getPropertyValue('background-color')).to.equal('rgb(255, 0, 0)');
-    });
-
-    it('should map the latest props combination classkey to the component', function test() {
-      if (/jsdom/.test(window.navigator.userAgent)) {
-        // see https://github.com/jsdom/jsdom/issues/2953
-        this.skip();
-      }
-
-      const theme = createMuiTheme({
-        variants: {
-          MuiButton: [
-            {
-              props: { variant: 'test' },
-              styles: { backgroundColor: 'rgb(255, 0, 0)' },
-            },
-            {
-              props: { variant: 'test', size: 'large', color: 'primary' },
-              styles: { backgroundColor: 'rgb(0, 255, 0)' },
-            },
-          ],
-        },
-      });
-
-      render(<WrappedComponent theme={theme} variant="test" size="large" color="primary" />);
-
-      const style = window.getComputedStyle(screen.getByTestId('component'));
-      expect(style.getPropertyValue('background-color')).to.equal('rgb(0, 255, 0)');
-    });
-
-    it('should not add classKey if not all props match', function test() {
-      if (/jsdom/.test(window.navigator.userAgent)) {
-        // see https://github.com/jsdom/jsdom/issues/2953
-        this.skip();
-      }
-
-      const theme = createMuiTheme({
-        variants: {
-          MuiButton: [
-            {
-              props: { variant: 'test' },
-              styles: { backgroundColor: 'rgb(255, 0, 0)' },
-            },
-            {
-              props: { variant: 'test', size: 'large' },
-              styles: { backgroundColor: 'rgb(0, 255, 0)' },
-            },
-          ],
-        },
-      });
-
-      render(<WrappedComponent theme={theme} size="large" />);
-
-      const style = window.getComputedStyle(screen.getByTestId('component'));
-      expect(style.getPropertyValue('background-color')).not.to.equal('rgb(0, 255, 0)');
-    });
-
-    it('should consider default props when matching the props', function test() {
-      if (/jsdom/.test(window.navigator.userAgent)) {
-        // see https://github.com/jsdom/jsdom/issues/2953
-        this.skip();
-      }
-
-      const theme = createMuiTheme({
-        variants: {
-          MuiButton: [
-            {
-              props: { variant: 'test' },
-              styles: { backgroundColor: 'rgb(0, 0, 0)' },
-            },
-            {
-              props: { variant: 'test', color: 'primary', size: 'large' },
-              styles: { backgroundColor: 'rgb(255, 0, 0)' },
-            },
-          ],
-        },
-      });
-
-      render(<WrappedComponent theme={theme} variant="test" size="large" />);
-
-      const style = window.getComputedStyle(screen.getByTestId('component'));
-      expect(style.getPropertyValue('background-color')).to.equal('rgb(255, 0, 0)');
-    });
-
-    it('should warn if the used variant is not defined in the theme', function test() {
-      const theme = createMuiTheme({
-        variants: {
-          MuiButton: [
-            {
-              props: { variant: 'test1' },
-              styles: { backgroundColor: 'rgb(255, 0, 0)' },
-            },
-          ],
-        },
-      });
-
-      expect(() => mount(<WrappedComponent theme={theme} variant="test" />)).toErrorDev([
-        // strict mode renders twice
-        [
-          `Material-UI: You are using a variant value \`test\` for which you didn't define styles.`,
-          `Please create a new variant matcher in your theme for this variant. To learn more about matchers visit https://material-ui.com/customization/components/#adding-new-component-variants.`,
-        ].join('\n'),
-        [
-          `Material-UI: You are using a variant value \`test\` for which you didn't define styles.`,
-          `Please create a new variant matcher in your theme for this variant. To learn more about matchers visit https://material-ui.com/customization/components/#adding-new-component-variants.`,
-        ].join('\n'),
-      ]);
     });
   });
 });
